+++ src/initial-submission/components/FileDetailsForm.tsx
@@ -108,7 +108,6 @@ const FileDetailsForm = ({ initialValues }: Props): JSX.Element => {
                 },
             };
             setSupportingFilesStatus(fileStates);
-            console.log('File States', fileStates);
         };
     }
 
@@ -119,11 +118,9 @@ const FileDetailsForm = ({ initialValues }: Props): JSX.Element => {
         const iterator = fileUploadInitializer(filesToStore);
 
         const loop = (result: IteratorResult<{ uploadPromise: {}; itemId: string }>, fileStates: FileState[]): void => {
-            console.log('going loopy....');
             if (result.done) {
                 setSupportingUploadDisabled(false);
             } else {
-                console.log(result.value);
                 uploadSupportingFile(result.value.uploadPromise)
                     .then(setSupportingFileState(result, fileStates))
                     .then(() => loop(iterator.next(), fileStates));
@@ -236,9 +233,6 @@ const FileDetailsForm = ({ initialValues }: Props): JSX.Element => {
     };
 
     useAutoSave(onSave, [coverLetter]);
-    useEffect(() => {
-        console.log('changed files status', supportingFilesStatus);
-    }, [supportingFilesStatus]);
 
     return (
         <div>
